remove unused files various minor fixes add sanitizer configuration fix install target for MSVC

+++ include/carve/polyline_iter.hpp
@@ -77,7 +77,7 @@ struct polyline_vertex_iter
 
 	Vertex* operator*() const
 	{
-		CARVE_ASSERT(idx >= 0 && idx < base->vertexCount());
+		CARVE_ASSERT(idx >= 0 && static_cast<size_t>(idx) < base->vertexCount());
 		return base->vertex((size_t)idx);
 	}
 };
@@ -163,7 +163,7 @@ struct polyline_vertex_const_iter
 
 	const Vertex* operator*() const
 	{
-		CARVE_ASSERT(idx >= 0 && idx < base->vertexCount());
+		CARVE_ASSERT(idx >= 0 && static_cast<size_t>(idx) < base->vertexCount());
 		return base->vertex((size_t)idx);
 	}
 };
@@ -258,7 +258,7 @@ struct polyline_edge_iter
 
 	PolylineEdge* operator*() const
 	{
-		CARVE_ASSERT(idx >= 0 && idx < base->edgeCount());
+		CARVE_ASSERT(idx >= 0 && static_cast<size_t>(idx) < base->edgeCount());
 		return base->edge((size_t)idx);
 	}
 };
@@ -344,7 +344,7 @@ struct polyline_edge_const_iter
 
 	const PolylineEdge* operator*() const
 	{
-		CARVE_ASSERT(idx >= 0 && idx < base->edgeCount());
+		CARVE_ASSERT(idx >= 0 && static_cast<size_t>(idx) < base->edgeCount());
 		return base->edge((size_t)idx);
 	}
 };
